8369258: C2: enable ReassociateInvariants for all loop types
