+++ test/hotspot/jtreg/compiler/loopopts/superword/TestMemorySegment_ReassociateInvariants2.java
@@ -31,22 +31,23 @@ import compiler.lib.ir_framework.*;
 
 /*
  * @test
- * @bug 8324751
- * @summary Reported issue: JDK-8360204: C2 SuperWord: missing RCE with MemorySegment.getAtIndex
+ * @bug 8324751 8369258
+ * @summary Reported issue: JDK-8365982: C2 SuperWord: missing RCE / strange Multiversioning with MemorySegment.set
  *          The examples are generated from TestAliasingFuzzer.java
  *          So if you see something change here, you may want to investigate if we
  *          can also tighten up the IR rules there.
  * @library /test/lib /
- * @run driver compiler.loopopts.superword.TestMemorySegment_8360204
+ * @run driver compiler.loopopts.superword.TestMemorySegment_ReassociateInvariants2
  */
 
-public class TestMemorySegment_8360204 {
+public class TestMemorySegment_ReassociateInvariants2 {
 
-    public static MemorySegment a = Arena.ofAuto().allocate(10_000);
-    public static MemorySegment b = Arena.ofAuto().allocate(10_000);
+    public static MemorySegment a = MemorySegment.ofArray(new short[100_000]);
+    public static MemorySegment b = MemorySegment.ofArray(new short[100_000]);
 
-    private static long invar0_1159 = 0;
-    private static long invar1_1159 = 0;
+    private static long invar0_853 = 0;
+    private static long invar1_853 = 0;
+    private static long invar2_853 = 0;
 
     public static void main(String[] args) {
         TestFramework f = new TestFramework();
@@ -60,31 +61,30 @@ public class TestMemorySegment_8360204 {
 
     @Setup
     static Object[] setup() {
-        return new Object[] { a, -19125L, b, 71734L + 2_000L, 0, 1_000 };
+        return new Object[] { a, -50_000, b, -30_000, 0, 10_000 };
     }
 
     @Test
     @Arguments(setup = "setup")
-    @IR(counts = {IRNode.LOAD_VECTOR_I, "= 0",
-                  IRNode.STORE_VECTOR,  "= 0",
-                  ".*multiversion.*",   "> 0"}, // Sadly, we now multiversion
+    @IR(counts = {IRNode.STORE_VECTOR, "> 0",
+                  IRNode.REPLICATE_S,  "> 0",
+                  ".*multiversion.*",  "= 0"}, // Good: The AutoVectorization predicate suffices.
         phase = CompilePhase.PRINT_IDEAL,
         applyIfPlatform = {"64-bit", "true"},
-        applyIf = {"AlignVector", "false"},
+        applyIfAnd = {"AlignVector", "false", "ShortRunningLongLoop", "false"},
         applyIfCPUFeatureOr = {"avx", "true", "asimd", "true"})
-    // There is no aliasing, so we should compile without multiversioning.
-    // But currently, there seems to be some issue with RCE, we peel and lose the predicate.
-    // Then we multiversion.
-    // We could imagine that this would eventually vectorize, but since one counts up, and the other down,
-    // we would have to implement shuffle first.
     //
-    // If you see this IR rule fail: investigate JDK-8360204, possibly close it and fix this IR rule!
-    // Also: consider renaming the file to something more descriptive: what have you fixed with this?
-    // And: you may now be able to tighten IR rules in TestAliasingFuzzer.java
+    @IR(counts = {IRNode.STORE_VECTOR, "> 0",
+                  IRNode.REPLICATE_S,  "> 0",
+                  ".*multiversion.*",  "= 0"},
+        phase = CompilePhase.PRINT_IDEAL,
+        applyIfPlatform = {"64-bit", "true"},
+        applyIfAnd = {"AlignVector", "false", "ShortRunningLongLoop", "true"},
+        applyIfCPUFeatureOr = {"avx", "true", "asimd", "true"})
     public static void test(MemorySegment container_0, long invar0_0, MemorySegment container_1, long invar0_1, long ivLo, long ivHi) {
-        for (long i = ivLo; i < ivHi; i+=1) {
-            var v = container_0.getAtIndex(ValueLayout.JAVA_INT_UNALIGNED, 19125L + 1L * i + 1L * invar0_0 + 0L * invar0_1159 + 1L * invar1_1159);
-            container_1.setAtIndex(ValueLayout.JAVA_INT_UNALIGNED, -71734L + -1L * i + 1L * invar0_1 + 1L * invar0_1159 + 0L * invar1_1159, v);
+        for (long i = ivHi-1; i >= ivLo; i-=1) {
+            container_0.set(ValueLayout.JAVA_CHAR_UNALIGNED, -47143L + -2L * i + -2L * invar0_0 + -1L * invar0_853 + -1L * invar1_853 + 0L * invar2_853, (char)0x0102030405060708L);
+            container_1.set(ValueLayout.JAVA_CHAR_UNALIGNED, 74770L + 2L * i + 2L * invar0_1 + 0L * invar0_853 + 0L * invar1_853 + 0L * invar2_853, (char)0x1112131415161718L);
         }
     }
 }
